mini_projects/stock_synchronizer/requirements.txt
PyYAML==3.12
ujson==1.35
